Hacky link replacement for non-markdown pages.
This commit adds a post-rendering hook for pages that replaces any links
to what appear to be markdown pages with what is probably the link to
the real page.

When I say this is hacky, I really mean it!

But it does the job for my homepage timeline and archive pages, and
that's what matters.